+++ boofuzz/sessions.py
@@ -12,6 +12,7 @@ import time
 import traceback
 import warnings
 import zlib
+from .exception import BoofuzzFailure
 from builtins import input
 from io import open
 
@@ -1190,19 +1191,19 @@ class Session(pgraph.Graph):
             if self._ignore_connection_reset:
                 self._fuzz_data_logger.log_info(constants.ERR_CONN_RESET)
             else:
-                self._fuzz_data_logger.log_fail(constants.ERR_CONN_RESET)
+                raise BoofuzzFailure(message=constants.ERR_CONN_RESET)
         except exception.BoofuzzTargetConnectionAborted as e:
             # TODO: Switch _ignore_connection_aborted for _ignore_transmission_error, or provide retry mechanism
             msg = constants.ERR_CONN_ABORTED.format(socket_errno=e.socket_errno, socket_errmsg=e.socket_errmsg)
             if self._ignore_connection_aborted:
                 self._fuzz_data_logger.log_info(msg)
             else:
-                self._fuzz_data_logger.log_fail(msg)
+                raise BoofuzzFailure(msg)
         except exception.BoofuzzSSLError as e:
             if self._ignore_connection_ssl_errors:
                 self._fuzz_data_logger.log_info(str(e))
             else:
-                self._fuzz_data_logger.log_fail(str(e))
+                raise BoofuzzFailure(message=str(e))
 
         try:  # recv
             if self._receive_data_after_each_request:
@@ -1212,25 +1213,25 @@ class Session(pgraph.Graph):
                     self._fuzz_data_logger.log_check("Verify some data was received from the target.")
                     if not self.last_recv:
                         # Assume a crash?
-                        self._fuzz_data_logger.log_fail("Nothing received from target.")
+                        raise BoofuzzFailure(message="Nothing received from target.")
                     else:
                         self._fuzz_data_logger.log_pass("Some data received from target.")
         except exception.BoofuzzTargetConnectionReset:
             if self._check_data_received_each_request:
-                self._fuzz_data_logger.log_fail(constants.ERR_CONN_RESET)
+                raise BoofuzzFailure(message=constants.ERR_CONN_RESET)
             else:
                 self._fuzz_data_logger.log_info(constants.ERR_CONN_RESET)
         except exception.BoofuzzTargetConnectionAborted as e:
             msg = constants.ERR_CONN_ABORTED.format(socket_errno=e.socket_errno, socket_errmsg=e.socket_errmsg)
             if self._check_data_received_each_request:
-                self._fuzz_data_logger.log_fail(msg)
+                raise BoofuzzFailure(msg)
             else:
                 self._fuzz_data_logger.log_info(msg)
         except exception.BoofuzzSSLError as e:
             if self._ignore_connection_ssl_errors:
                 self._fuzz_data_logger.log_info(str(e))
             else:
-                self._fuzz_data_logger.log_fail(str(e))
+                raise BoofuzzFailure(str(e))
 
     def transmit_fuzz(self, sock, node, edge, callback_data):
         """Render and transmit a fuzzed node, process callbacks accordingly.
@@ -1253,31 +1254,31 @@ class Session(pgraph.Graph):
             if self._ignore_connection_issues_when_sending_fuzz_data:
                 self._fuzz_data_logger.log_info(constants.ERR_CONN_RESET)
             else:
-                self._fuzz_data_logger.log_fail(constants.ERR_CONN_RESET)
+                raise BoofuzzFailure(message=constants.ERR_CONN_RESET)
         except exception.BoofuzzTargetConnectionAborted as e:
             msg = constants.ERR_CONN_ABORTED.format(socket_errno=e.socket_errno, socket_errmsg=e.socket_errmsg)
             if self._ignore_connection_issues_when_sending_fuzz_data:
                 self._fuzz_data_logger.log_info(msg)
             else:
-                self._fuzz_data_logger.log_fail(msg)
+                raise BoofuzzFailure(msg)
         except exception.BoofuzzSSLError as e:
             if self._ignore_connection_ssl_errors:
                 self._fuzz_data_logger.log_info(str(e))
             else:
-                self._fuzz_data_logger.log_fail(str(e))
+                raise BoofuzzFailure(str(e))
 
         try:  # recv
             if self._receive_data_after_fuzz:
                 self.last_recv = self.targets[0].recv()
         except exception.BoofuzzTargetConnectionReset:
             if self._check_data_received_each_request:
-                self._fuzz_data_logger.log_fail(constants.ERR_CONN_RESET)
+                raise BoofuzzFailure(message=constants.ERR_CONN_RESET)
             else:
                 self._fuzz_data_logger.log_info(constants.ERR_CONN_RESET)
         except exception.BoofuzzTargetConnectionAborted as e:
             msg = constants.ERR_CONN_ABORTED.format(socket_errno=e.socket_errno, socket_errmsg=e.socket_errmsg)
             if self._check_data_received_each_request:
-                self._fuzz_data_logger.log_fail(msg)
+                raise BoofuzzFailure(msg)
             else:
                 self._fuzz_data_logger.log_info(msg)
             pass
@@ -1285,7 +1286,7 @@ class Session(pgraph.Graph):
             if self._ignore_connection_ssl_errors:
                 self._fuzz_data_logger.log_info(str(e))
             else:
-                self._fuzz_data_logger.log_fail(str(e))
+                raise BoofuzzFailure(str(e))
 
     def build_webapp_thread(self, port=constants.DEFAULT_WEB_UI_PORT):
         app.session = self
@@ -1574,6 +1575,10 @@ class Session(pgraph.Graph):
             if self.sleep_time > 0:
                 self._fuzz_data_logger.open_test_step("Sleep between tests.")
                 self._sleep(self.sleep_time)
+        except BoofuzzFailure as e:
+            self._fuzz_data_logger.log_fail(e.message)
+            if not self._check_for_passively_detected_failures(target=target):
+                self._check_for_passively_detected_failures(target=target)
         finally:
             self._process_failures(target=target)
             self._fuzz_data_logger.close_test_case()